pyproject.toml
[project]
name = "eduagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "ffmpeg-python>=0.2.0",
    "flask>=3.0.0",
    "gradio>=5.43.1",
    "litellm>=1.75.9",
    "manim>=0.19.0",
    "networkx>=3.5",
    "openai>=1.0.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydot>=4.0.1",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.0",
    "uvicorn[standard]>=0.35.0",
]
